move Filter component to /components
