+++ src/components/Filter/Chip/Chip.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .button_element {
     position: relative;
