+++ src/components/Filter/OptionElement/OptionElement.module.scss
@@ -1,4 +1,4 @@
-@import '../../../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .option {
     display: flex;
